# Conflicts:
#	src/sources/aws_kinesis_firehose/filters.rs
#	src/sources/aws_kinesis_firehose/mod.rs

+++ lib/lookup/src/lookup_v2/jit.rs
@@ -286,7 +286,7 @@ impl<'a> Iterator for JitValuePathIter<'a> {
                                 self.coalesce_count += 1;
                                 (
                                     Some(Some(BorrowedSegment::CoalesceField(Cow::Borrowed(
-                                        &self.path[(start as usize)..(end as usize)],
+                                        &self.path[start..end],
                                     )))),
                                     JitState::CoalesceStart,
                                 )
@@ -298,7 +298,7 @@ impl<'a> Iterator for JitValuePathIter<'a> {
                                     self.coalesce_count = 0;
                                     (
                                         Some(Some(BorrowedSegment::CoalesceEnd(Cow::Borrowed(
-                                            &self.path[(start as usize)..(end as usize)],
+                                            &self.path[start..end],
                                         )))),
                                         JitState::Continue,
                                     )
